Attendance

+++ src/main/java/synergyhubback/attendance/dto/request/DayOffBalanceRequest.java
@@ -15,17 +15,17 @@ import java.time.LocalDate;
 public class DayOffBalanceRequest {
 
     private int dbCode;                 //휴가관리코드(pk)
-    private int granted;                //부여수
-    private int remaining;              //잔여수
-    private int dbUsed;                 //사용수
+    private Double granted;                //부여수
+    private Double remaining;              //잔여수
+    private Double dbUsed;                 //사용수
     private Employee employee;          //사원코드
     private LocalDate dbInsertDate;     //부여일자
 
     @JsonCreator
     public DayOffBalanceRequest(@JsonProperty("dbCode") int dbCode,
-                                 @JsonProperty("granted") int granted,
-                                 @JsonProperty("remaining") int remaining,
-                                 @JsonProperty("dbUsed") int dbUsed,
+                                 @JsonProperty("granted") Double granted,
+                                 @JsonProperty("remaining") Double remaining,
+                                 @JsonProperty("dbUsed") Double dbUsed,
                                  @JsonProperty("employee") Employee employee,
                                  @JsonProperty("dbInsertDate") LocalDate dbInsertDate) {
         this.dbCode = dbCode;
